alinea d-k

+++ TP2/ADO.NET/TP2_ADO.NET/Entidades/Inscricao.cs
@@ -13,7 +13,7 @@ namespace Entidades
         public string siglaSemestreLetivo { get; set; }
         public int ano { get; set; }
         public string siglaUC { get; set; }
-        public float nota { get; set; }
+        public double nota { get; set; }
 
     }
 }
